Refactor save into workspace plugin
