+++ rmf_site_editor/src/lib.rs
@@ -15,8 +15,6 @@ pub mod animate;
 pub mod keyboard;
 use keyboard::*;
 
-pub mod save;
-use save::*;
 pub mod widgets;
 use widgets::{menu_bar::MenuPluginManager, *};
 
@@ -202,7 +200,6 @@ impl Plugin for SiteEditor {
                 AabbUpdatePlugin,
                 EguiPlugin,
                 KeyboardInputPlugin,
-                SavePlugin,
                 SdfPlugin,
                 MainMenuPlugin,
                 WorkcellEditorPlugin,
