+++ rmf_site_editor/src/workspace.rs
@@ -21,8 +21,8 @@ use std::collections::BTreeMap;
 use std::path::PathBuf;
 
 use crate::interaction::InteractionState;
-use crate::site::LoadSite;
-use crate::workcell::LoadWorkcell;
+use crate::site::{DefaultFile, LoadSite, SaveSite};
+use crate::workcell::{LoadWorkcell, SaveWorkcell};
 use crate::AppState;
 use rmf_site_format::legacy::building_map::BuildingMap;
 use rmf_site_format::{Level, NameOfSite, Site, SiteProperties, Workcell};
@@ -107,6 +107,81 @@ impl Default for LoadWorkspaceChannels {
     }
 }
 
+#[derive(Event)]
+pub struct SaveWorkspace {
+    /// If specified workspace will be saved to requested file, otherwise the default file
+    pub destination: SaveWorkspaceDestination,
+    /// If specified the workspace will be exported to a specific format
+    pub format: ExportFormat,
+}
+
+impl SaveWorkspace {
+    pub fn new() -> Self {
+        Self {
+            destination: SaveWorkspaceDestination::default(),
+            format: ExportFormat::default(),
+        }
+    }
+
+    pub fn to_default_file(mut self) -> Self {
+        self.destination = SaveWorkspaceDestination::DefaultFile;
+        self
+    }
+
+    pub fn to_dialog(mut self) -> Self {
+        self.destination = SaveWorkspaceDestination::Dialog;
+        self
+    }
+
+    pub fn to_path(mut self, path: &PathBuf) -> Self {
+        self.destination = SaveWorkspaceDestination::Path(path.clone());
+        self
+    }
+
+    pub fn to_urdf(mut self) -> Self {
+        self.format = ExportFormat::Urdf;
+        self
+    }
+}
+
+#[derive(Default, Debug, Clone)]
+pub enum SaveWorkspaceDestination {
+    #[default]
+    DefaultFile,
+    Dialog,
+    Path(PathBuf),
+}
+
+#[derive(Clone, Default, Debug)]
+pub enum ExportFormat {
+    #[default]
+    Default,
+    Urdf,
+}
+
+/// Event used in channels to communicate the file handle that was chosen by the user.
+#[derive(Debug)]
+pub struct SaveWorkspaceFile {
+    path: PathBuf,
+    format: ExportFormat,
+    root: Entity,
+}
+
+/// Use a channel since file dialogs are async and channel senders can be cloned and passed into an
+/// async block.
+#[derive(Debug, Resource)]
+pub struct SaveWorkspaceChannels {
+    pub sender: Sender<SaveWorkspaceFile>,
+    pub receiver: Receiver<SaveWorkspaceFile>,
+}
+
+impl Default for SaveWorkspaceChannels {
+    fn default() -> Self {
+        let (sender, receiver) = crossbeam_channel::unbounded();
+        Self { sender, receiver }
+    }
+}
+
 /// Used to keep track of visibility when switching workspace
 #[derive(Debug, Default, Resource)]
 pub struct RecallWorkspace(Option<Entity>);
@@ -125,8 +200,10 @@ impl Plugin for WorkspacePlugin {
         app.add_event::<ChangeCurrentWorkspace>()
             .add_event::<CreateNewWorkspace>()
             .add_event::<LoadWorkspace>()
+            .add_event::<SaveWorkspace>()
             .init_resource::<CurrentWorkspace>()
             .init_resource::<RecallWorkspace>()
+            .init_resource::<SaveWorkspaceChannels>()
             .init_resource::<LoadWorkspaceChannels>()
             .add_systems(
                 Update,
@@ -135,8 +212,11 @@ impl Plugin for WorkspacePlugin {
                     workspace_file_load_complete,
                     sync_workspace_visibility,
                     dispatch_load_workspace_events,
+                    workspace_file_save_complete,
                 ),
             );
+        #[cfg(not(target_arch = "wasm32"))]
+        app.add_systems(Update, dispatch_save_workspace_events);
     }
 }
 
@@ -175,7 +255,7 @@ pub fn dispatch_new_workspace_events(
 }
 
 pub fn dispatch_load_workspace_events(
-    mut load_channels: ResMut<LoadWorkspaceChannels>,
+    load_channels: Res<LoadWorkspaceChannels>,
     mut load_workspace: EventReader<LoadWorkspace>,
 ) {
     if let Some(cmd) = load_workspace.iter().last() {
@@ -225,7 +305,7 @@ fn workspace_file_load_complete(
     mut interaction_state: ResMut<NextState<InteractionState>>,
     mut load_site: EventWriter<LoadSite>,
     mut load_workcell: EventWriter<LoadWorkcell>,
-    mut load_channels: ResMut<LoadWorkspaceChannels>,
+    load_channels: Res<LoadWorkspaceChannels>,
 ) {
     if let Ok(result) = load_channels.receiver.try_recv() {
         let LoadWorkspaceFile(default_file, data) = result;
@@ -291,6 +371,94 @@ fn workspace_file_load_complete(
                     }
                 }
             }
+        }
+    }
+}
+
+// TODO(luca) implement this in wasm, it's supported since rfd version 0.12, however it requires
+// calling .write on the `FileHandle` object returned by the AsyncFileDialog. Such FileHandle is
+// not Send in wasm so it can't be sent to another thread through an event. We would need to
+// refactor saving to be fully done in the async task rather than send an event to have wasm saving.
+#[cfg(not(target_arch = "wasm32"))]
+fn dispatch_save_workspace_events(
+    mut save_events: EventReader<SaveWorkspace>,
+    save_channels: Res<SaveWorkspaceChannels>,
+    workspace: Res<CurrentWorkspace>,
+    default_files: Query<&DefaultFile>,
+) {
+    let spawn_dialog = |format: &ExportFormat, root| {
+        let sender = save_channels.sender.clone();
+        let format = format.clone();
+        AsyncComputeTaskPool::get()
+            .spawn(async move {
+                if let Some(file) = AsyncFileDialog::new().save_file().await {
+                    let path = file.path().to_path_buf();
+                    sender
+                        .send(SaveWorkspaceFile { path, format, root })
+                        .expect("Failed sending save event");
+                }
+            })
+            .detach();
+    };
+    for event in save_events.iter() {
+        if let Some(ws_root) = workspace.root {
+            match &event.destination {
+                SaveWorkspaceDestination::DefaultFile => {
+                    if let Some(file) = default_files.get(ws_root).ok().map(|f| f.0.clone()) {
+                        save_channels
+                            .sender
+                            .send(SaveWorkspaceFile {
+                                path: file,
+                                format: event.format.clone(),
+                                root: ws_root,
+                            })
+                            .expect("Failed sending save request");
+                    } else {
+                        spawn_dialog(&event.format, ws_root);
+                    }
+                }
+                SaveWorkspaceDestination::Dialog => spawn_dialog(&event.format, ws_root),
+                SaveWorkspaceDestination::Path(path) => {
+                    save_channels
+                        .sender
+                        .send(SaveWorkspaceFile {
+                            path: path.clone(),
+                            format: event.format.clone(),
+                            root: ws_root,
+                        })
+                        .expect("Failed sending save request");
+                }
+            }
+        } else {
+            warn!("Unable to save, no workspace loaded");
+            return;
+        }
+    }
+}
+
+/// Handles the file saving events
+fn workspace_file_save_complete(
+    app_state: Res<State<AppState>>,
+    mut save_site: EventWriter<SaveSite>,
+    mut save_workcell: EventWriter<SaveWorkcell>,
+    save_channels: Res<SaveWorkspaceChannels>,
+) {
+    if let Ok(result) = save_channels.receiver.try_recv() {
+        match app_state.get() {
+            AppState::WorkcellEditor => {
+                save_workcell.send(SaveWorkcell {
+                    root: result.root,
+                    to_file: result.path,
+                    format: result.format,
+                });
+            }
+            AppState::SiteEditor | AppState::SiteDrawingEditor | AppState::SiteVisualizer => {
+                save_site.send(SaveSite {
+                    site: result.root,
+                    to_file: result.path,
+                });
+            }
+            AppState::MainMenu => { /* Noop */ }
         }
     }
 }
